Handlin exceptions

+++ src/main/java/br/com/librumbr/services/BookService.java
@@ -1,5 +1,6 @@
 package br.com.librumbr.services;
 
+import br.com.librumbr.exceptions.BookNotFoundException;
 import br.com.librumbr.models.Author;
 import br.com.librumbr.models.Book;
 import br.com.librumbr.models.Category;
@@ -51,14 +52,14 @@ public class BookService {
                 .retrieve()
                 .onStatus(HttpStatusCode::is4xxClientError, response -> {
                     // Se for erro 404, ignora e tenta buscar local
-                    return Mono.error(new RuntimeException("Livro não encontrado na API"));
+                    return Mono.error(new BookNotFoundException("Livro não encontrado na API externa."));
                 })
                 .bodyToMono(BrasilApiResponseDTO.class)
                 .onErrorResume(ex -> {
                     // Busca no banco se falhar a chamada
                     Optional<Book> bookOptional = bookRepository.findByIsbn(isbn);
                     if (bookOptional.isEmpty()){
-                        return Mono.error(new RuntimeException("Livro não encontrado no banco de dados"));
+                        return Mono.error(new BookNotFoundException("Livro não encontrado no banco de dados"));
                     }
                     BrasilApiResponseDTO dto = ModelMapperUtil.parseObject(bookOptional.get(), BrasilApiResponseDTO.class);
                     return Mono.just(dto);
@@ -75,7 +76,7 @@ public class BookService {
 
     @Transactional
     public BookResponseDTO findBookById(int id) {
-        var book = bookRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+        var book = bookRepository.findById(id).orElseThrow(() -> new BookNotFoundException("Livro com ID " + id + " não encontrado"));
         return ModelMapperUtil.parseObject(book, BookResponseDTO.class);
     }
 
@@ -104,6 +105,9 @@ public class BookService {
 
     @Transactional
     public void deleteBook(int id) {
+        if (!bookRepository.existsById(id)) {
+            throw new BookNotFoundException("Livro com ID " + id + " não encontrado para exclusão.");
+        }
         bookRepository.deleteById(id);
     }
 
@@ -111,7 +115,7 @@ public class BookService {
     public void updateBook(int id, BookCreateDTO updatedBook) {
 
         Book oldBook = bookRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Book not found"));
+                .orElseThrow(() -> new BookNotFoundException("Livro não encontrado"));
 
         List<Author> authors = updatedBook.getAuthors().stream()
                 .map(authorName -> authorRepository.findByName(authorName)
